Splitted the code into new file and added the condition for admin usage only
